First code attempt at having configurable visualizers
