+++ data/config.json
@@ -9,5 +9,6 @@
   "night_lights": true,
   "night_populated_yellow": false,
   "night_category_proportion": 0.05,
-  "brightness_proportion": 1.0
+  "brightness_proportion": 1.0,
+  "visualizer": 0
 }
